% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enc.R
\name{enc}
\alias{enc}
\title{Encrypt a message}
\usage{
enc(pk, m)
}
\arguments{
\item{pk}{a public key for any scheme as generated by the \code{\link{keygen}}.
function.}

\item{m}{an integer to be encrypted.  Note that the permissable range of values
for \code{m} is dependent on the scheme and the parameters of the scheme.
\code{m} may even be resticted to as little as \{0,1\}, for example in binary
encryption schemes.}
}
\value{
A ciphertext under the encryption scheme, encrypted using the public key
provided.
}
\description{
This encrypts an integer message under one of the homomorphic encryption schemes
supported by this package.
}
\details{
The scheme to use is determined by the public key which contains the parameter
values for that scheme as part of it.

If a symmetric key scheme is being used, then the secret key should be provided
for the \code{pk} argument.
}
\examples{
p <- pars("FandV")
keys <- keygen(p)
ct <- enc(keys$pk, 1)
dec(keys$sk, ct)
}
\author{
Louis Aslett
}
\seealso{
\code{\link{keygen}} to create public/private keypairs;
\code{\link{dec}} to decrypt the ciphertext generated by this function.
}

