% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FandV.R
\name{FandV}
\alias{FandV}
\title{Fan and Vercauteren encryption scheme}
\usage{
p <- pars("FandV")
}
\description{
The Fan and Vercauteren scheme is implemented in this package.
}
\details{
Description of the scheme.
}
\examples{
# Benchmark the performance of the scheme
library(microbenchmark)
p <- pars("FandV")
microbenchmark({ keys <- keygen(p) }, unit="ms")
microbenchmark({ ct1 <- enc(keys$pk, 2) }, unit="ms")
ct2 <- enc(keys$pk, 3)
microbenchmark({ ct1 + ct2 }, unit="ms")
microbenchmark({ ct1 * ct2 }, unit="ms")
microbenchmark({ dec(keys$sk, ct1) }, unit="ms")
}

