% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keygen.R
\name{keygen}
\alias{keygen}
\title{Generate cryptographic keys}
\usage{
keygen(p)
}
\arguments{
\item{p}{a parameters object as produced by the \code{\link{pars}} function.}
}
\value{
A list object containing the keys will be returned

Depending on the scheme specified by the parameters object the exact structure
will vary.  For example, a symmetric key scheme will only return a private key,
whereas a public key scheme will return a public and private keypair.
Certain schemes may include additional keys for operations such as
relinearisation or bootstrapping.  The keys will be named \code{pk} (public),
\code{sk} (private), \code{rlk} (relinearisation), \code{bk} (bootstrapping)
in the list object if they are part of the scheme.
}
\description{
This function will generate cryptographic keys to enable encryption, decryption
and possibly other operations (such as relinearisation or bootstrapping) for
the homomorphic encryption schemes supported in this package.
}
\details{
The scheme to be used is determined by the type of the parameters object,
\code{p}.
}
\examples{
p <- pars("FandV")
keys <- keygen(p)

# Look at public key and encrypt
keys$pk
ct <- enc(keys$pk, 1)

# Look at private key and decrypt
keys$sk
dec(keys$sk, ct)

# Obviously a different private key won't be able to decrypt
keys2 <- keygen(p)
keys2$sk
dec(keys2$sk, ct)
}
\author{
Louis Aslett
}
\seealso{
\code{\link{pars}} for generating the parameters for a scheme; \code{\link{enc}}
for encrypting messages using the keys generated by this function.
}

