% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Setup encryption scheme parameters}
\usage{
pars(scheme, ...)
}
\arguments{
\item{scheme}{the scheme for which to create a parameter object.  Currently
only Fan and Vercauteren's scheme is supported by specifying \code{"FandV"}.}

\item{...}{pass the specific options for the chosen scheme as named arguments
to override any default values.  See the details section for options for
encryption schemes currently implemented.}
}
\description{
Use this function to create an encryption scheme parameters object.
}
\details{
Currently only the scheme of Fan and Vercauteren (\code{"FandV"}) is implemented.

For \code{"FandV"} you may specify:
\describe{
  \item{\code{d}}{power of the cyclotomic polynomial ring (default 4096);}
  \item{\code{sigma}}{the standard deviation of the discrete Gaussian used to
induce a distribution on the cyclotomic polynomial ring (default 16.0);}
  \item{\code{qpow}}{the power of 2 to use for the coefficient modulus (default 128);}
  \item{\code{t}}{the value to use for the message space modulus (default 32768).}
}

This function simply sets up the parameters which must be specified to use a
particular encryption scheme.  Using the scheme then requires generating
cryptographic keys.
}
\examples{
# Simplest example
p <- pars("FandV")
keys <- keygen(p)
ct <- enc(keys$pk, 1)
dec(keys$sk, ct)

# Change degree of cyclotomic polynomial used for ciphertext
p <- pars("FandV", d=2048)
keys <- keygen(p)
ct <- enc(keys$pk, 1)
dec(keys$sk, ct)
}
\author{
Louis Aslett
}
\references{
Fan, J., & Vercauteren, F. (2012). Somewhat Practical Fully Homomorphic
Encryption. IACR ePrint. Retrieved from \url{https://eprint.iacr.org/2012/144}
}
\seealso{
\code{\link{parsHelp}} for automatic assistance selecting parameters which
achieve a certain security level and multiplicative depth.

\code{\link{keygen}} to generate encryption keys using these parameters.
}

